Refactor adapter service into modular helpers
