+++ backend/services/adapters/__init__.py
@@ -1,0 +1,6 @@
+"""Adapter service package."""
+
+from .search import AdapterSearchResult
+from .service import AdapterService
+
+__all__ = ["AdapterService", "AdapterSearchResult"]
